firmware: arm_scmi/perf: add implementation of PERF_OPP_SET cmd
This patch adds required definitions and implementation of command
PERF_OPP_SET. Message id 0xF is allocated for this command.
The command allows to program tables for so-called 'traffic-cop'
residing in firmware.

The expected arguments/parameters for this command are:

u32 DVFS(performance) domain id;
u32 number of CPUs online (in this dvfs domain);
u32 maximum allowed OPP when this (see previous parameter) number of
CPUs are online;

Signed-off-by: Alexey Klimov <alexey.klimov@arm.com>