sched: Introduce energy models of CPUs
The energy consumption of each CPU in the system is modeled with a series
of values representing the dissipated power and compute capacity at each
available Operating Performance Point (OPP). In addition, the energy model
is provided with a simple representation of all frequency domains as
cpumasks, hence enabling the scheduler to be aware of dependencies
between CPUs. The new data structures holding these models and the
routines to populate them are stored in kernel/sched/energy.c.

As it appears that energy-aware scheduling really makes a difference on
heterogeneous systems (e.g. big.LITTLE platforms), it's utilisation is
kept limited to systems having:
   1. the SD_ASYM_CPUCAPACITY flag set;
   2. CPU frequency switching enabled;
   3. and available power estimates for the OPPs of all possible CPUs;

Moreover, the scheduler is notified of the energy model availability
using a static key hence minimizing the overhead on non energy-aware
systems.

Pending issues:
***************

 - !CONFIG_SMP is probably broken. We need add stubs to clean things up.

Not-signed-off-by: Quentin Perret <quentin.perret@arm.com>