sched: Introduce energy models of CPUs
The energy consumption of each CPU in the system is modeled with a list
of values representing its dissipated power and compute capacity at each
available Operating Performance Point (OPP). In addition, the energy model
is provided with a simple representation of all frequency domains as
cpumasks, hence enabling the scheduler to be aware of dependencies
between CPUs. The data required to build the energy model is provided by
the OPP library which enables an abstract view of the platform from the
scheduler. The new data structures holding these models and the routines
to populate them are stored in kernel/sched/energy.c.

As it appears that energy-aware scheduling really makes a difference on
heterogeneous systems (e.g. big.LITTLE platforms), its utilisation is
restricted to arm/arm64 systems having:
   1. the SD_ASYM_CPUCAPACITY flag set;
   2. CPU frequency switching enabled;
   3. available power estimates for the OPPs of all possible CPUs;

Moreover, the scheduler is notified of the energy model availability
using a static key in order to minimize the overhead on non-energy-aware
systems.

Cc: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Signed-off-by: Quentin Perret <quentin.perret@arm.com>

---
This patch depends on additional infrastructure being merged in the OPP
core. As this infrastructure can also be useful for other clients, the
related patches have been posted separately [1].

[1] https://marc.info/?l=linux-pm&m=151635516419249&w=2