BACKPORT: PM / OPP: Add an OPP power estimation helper
The dev_pm_opp data structure contains useful data to estimate power
consumption. In particular, when the u_amp field of the OPP's power supply
is known, the consumed power can be simply estimated as P = V * I.

This code provides a power estimation helper returning the OPP's power in
micro-watts when possible and 0 otherwise.

BACKPORT: OPP library in a different location

Change-Id: I7edb12217902d363eaf34e4dda43bd46ff139da7
Signed-off-by: Quentin Perret <quentin.perret@arm.com>