sched/fair: Introduce an energy estimation helper function
In preparation for the definition of an energy-aware wakeup path, a
helper function is provided to estimate the consequence on system energy
when a specific task wakes-up on a specific CPU. compute_energy()
estimates the OPPs to be reached by all frequency domains and estimates
the consumption of each online CPU according to its energy model.

More precisely, the capactity state reached by CPUs belonging to a
frequency domain is estimated using the max utilization of those CPUs.
Then, the utilization of each CPU is normalized to its expected capacity
state to estimate its busy time ratio. Finally the energy of the CPU is
estimated as the product of its normalized util (which is to be seen as
time) and its capacity-state power.

Pending issues:
 - I need to document compute_energy() properly, especially to state
   that it's safe to use it only if sched_energy_present if true.
 - I assume that all CPUs in a frequency domain share their capacity
   states. That helps a lot keeping things simple. However, this might
   not be true for all platforms (ie DynamIQ) so we have to discuss that
   topic.