sched/fair: Select an energy-efficient CPU on task wake-up
Whenever an energy model is available, let's enable an energy-aware
placement of tasks on wake-up. The eligible CPUs to be used in the
energy-aware wakeup path are restricted to the highest non-overutilized
sched_domain hence enabling energy aware placement only for lightly
loaded scenarios.

The selection of the most energy-efficient CPU is achieved by estimating
the system-level energy resulting from the placement of the task on each
candidate. The best CPU energy-wise is selected only if it saves a large
enough amount of energy with respect to prev_cpu. This is done to
prevent costly migrations for non-significant energy savings.

This brute force approach cannot scale to platforms with numerous CPUs.
If the scope of usage of the energy-aware wakeup path is broadened,
a better heuristic will have to be written.

Cc: Ingo Molnar <mingo@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Signed-off-by: Quentin Perret <quentin.perret@arm.com>