sched/fair: Select an energy-efficient CPU on task wake-up
Whenever an energy model is available, let's enable an energy-aware
placement of tasks on wake-up. The eligible CPUs to be used in the
energy-aware wakeup path are restricted to the highest non-overutilized
sched_domain hence enabling energy aware placement only for lightly
loaded scenarios.

The selection of the most energy-efficient CPU for executing a task is
achieved by estimating the system-level energy assuming the placement of
the task on each eligible CPU. The best CPU energy-wise is chosen only
if it saves enough energy with respect to prev_cpu, hence preventing
costly migrations for non-significant energy savings.

Although this brute force approach does not scale to platforms with
numerous CPUs, the overhead on mobile platforms (which are the preferred
targets for energy aware scheduling right now) remains acceptable.

Signed-off-by: Quentin Perret <quentin.perret@arm.com>