sched/fair: Select an energy-efficient CPU on task wake-up
Whenever an energy model is available, waking tasks are re-routed into a
new energy-aware placement algorithm. The eligible CPUs to be used in the
energy-aware wakeup path are restricted to the highest non-overutilized
sched_domain containing prev_cpu and this_cpu. Whenever no such scheduling
domain is found, the tasks go through the usual wake-up path, hence
enabling energy-aware placement only in lightly utilized scenarios.

The selection of the most energy-efficient CPU for a task is achieved by
estimating the system-level energy resulting from the placement of the
task on each and every possible candidate. The best CPU energy-wise is
then selected if it saves a large enough amount of energy with respect to
prev_cpu.

Although it has already shown significant benefits on some existing targets,
this brute force approach clearly cannot scale to platforms with numerous
CPUs. This patch is an attempt to do something useful as writing a fast
heuristic that performs reasonably well on a broad spectrum of architectures
isn't an easy task. As a consequence, for now, the scope of usability of the
energy-aware wake-up path is strictly restricted to systems having the
SD_ASYM_CPUCAPACITY flag, which not only show the most promising
opportunities for improving the energy consumption, but also often feature
a limited number of cores (at the time of writing).

Cc: Ingo Molnar <mingo@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Signed-off-by: Quentin Perret <quentin.perret@arm.com>