perf: Optimize ctx_sched_out()
When an event group contains more events than can be scheduled on the
hardware, iterateing the full event group for ctx_sched_out is a waste
of time.

Keep track of the events that got programmed on the hardware, such
that we can iterate this smaller list in order to schedule them out.

Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>