CONFIG: remove DEVFREQ_THERMAL from defconfig
With latest android kernel v4.9, a timer fire every 400us.
This timer check if it can power off the gpu which never happen even
during idle screen off use case. Until this is fixed, we should disable
CONFIG_DEVRFEQ_THERMAL that is used to register the gpu as a thermal
cooling device.
The power impact can be up to 15%

Bug: https://bugs.96boards.org/show_bug.cgi?id=652

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Reported-by: Vincent Guittot <vincent.guittot@linaro.org>