tracing: Fix trace_preempt_{on,off} for !CONFIG_DEBUG_PREEMPT builds
The new preempt enable/disable events introduced by:

 d59158162 tracing: Add support for preempt and irq enable/disable events

are defined only for CONFIG_DEBUG_PREEMPT kernels when the
CONFIG_PREEMPTIRQ_EVENTS support, introduced by the above patch, is
enabled.

These events are generated within by trace_preempt_{on,off} calls,
which are part of the the "Preemption-off Latency Tracer"
(CONFIG_PREEMPT_TRACER).

Currently these calls are (correctly) generated on CONFIG_PREEMPT_TRACER
even when the kernel is !CONFIG_DEBUG_PREEMPT configured. Thus leading
to an undefined reference to the new trace events.

Let's ensure that we always have an (eventually) empty definition of the
events for CONFIG_PREEMPTIRQ_EVENTS kernels.

This patch will ensure that the additional preempt enabled/disable
events are generated only when we have both:
 CONFIG_PREEMPT_TRACER && CONFIG_DEBUG_PREEMPT

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Reviewed-by: Dietmar Eggemann <dietmar.eggemann@arm.com>
Reviewed-by: Alessio Balsini <alessio.balsini@arm.com>
Cc: Steven Rostedt <rostedt@goodmis.org>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: Joel Fernandes <joelaf@google.com>
Cc: Linux-kernel@vger.kernel.org

---
Fix kernel compilation for this configuration:
 !CONFIG_DEBUG_PREEMPT && CONFIG_PREEMPT_TRACER

To reproduce, e.g.
 1. make defconfig
 2. add:
    CONFIG_PREEMPTIRQ_EVENTS=y
    CONFIG_PREEMPT_TRACER=y
 3. make kernel/trace/trace_irqsoff.o

For sake of consistency, this patch adds a default empty definition of
all the new trace events whenever CONFIG_PREEMPTIRQ_EVENTS is enabled
but the specific event's requirements are not satisfied.
This is complementary to the default empty definitions provided
for the !CONFIG_PREEMPTIRQ_EVENTS case.