x86/ldt: Rename ldt_struct->entries member
To support user shared LDT entry mappings it's required to change the LDT
related code so that the kernel side only references the real page mapping
of the LDT. When the LDT is loaded then the entries are alias mapped in the
per cpu fixmap. To catch all users rename ldt_struct->entries and fix them
up.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>