perf: implement pmu perf_kprobe
A new pmu, perf_kprobe is added. Based attr from perf_event_open(),
perf_kprobe creates a kprobe (or kretprobe) for the perf_event. This
kprobe is private to this perf_event, and thus not added to global
lists, and not available in tracefs.

Two functions, create_local_trace_kprobe() and
destroy_local_trace_kprobe()  are added to created and destroy these
local trace_kprobe.

Cc: <rostedt@goodmis.org>
Cc: <davem@davemloft.net>
Cc: <mingo@redhat.com>
Cc: <daniel@iogearbox.net>
Cc: CC: <kernel-team@fb.com>
Signed-off-by: Song Liu <songliubraving@fb.com>
Reviewed-by: Yonghong Song <yhs@fb.com>
Reviewed-by: Josef Bacik <jbacik@fb.com>
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Link: http://lkml.kernel.org/r/20171206224518.3598254-6-songliubraving@fb.com