sched/fair: select the most energy-efficient CPU candidate on wake-up
The current implementation of the energy-aware wake-up path relies on
find_best_target() to select an ordered list of CPU candidates for task
placement. The first candidate of the list saving energy is then chosen,
disregarding all the others to avoid the overhead of an expensive
energy_diff.

With the recent refactoring of select_energy_cpu_idx(), the cost of
exploring multiple CPUs has been reduced, hence offering the opportunity
to select the most energy-efficient candidate at a lower cost. This commit
seizes this opportunity by changing select_energy_cpu_idx()'s behaviour as
to ignore the order in the list of CPUs returned by find_best_target() and
to pick the one enabling the best energy savings.

Change-Id: I0cb833bfec1a4a053eddaff1652c0b6cad554f97
Signed-off-by: Quentin Perret <quentin.perret@arm.com>