cpufreq/schedutil: Rewrite CPUFREQ_RT support
Instead of trying to duplicate scheduler state to track if an RT task
is running, directly use the scheduler runqueue state for it.

This vastly simplifies things and fixes a number of bugs related to
sugov and the scheduler getting out of sync wrt this state.

Also (maybe split out):

 - update the remote cfs/dl state when iterating the shared mask
 - apply io-boost to the local util before finding the shared max

Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>