sched: Rename root_domain->overload to should_idle_balance
The name "overload" is not very explicit, especially since it doesn't
use any concept of "load" coming from load-tracking signals. For now it
simply tracks if any of the CPUs in the sched_groups have more than one
runnable task, and is then used to decide whether idle balance should
be performed.

As such, this commit changes that flag's name to 'should_idle_balance',
which makes its role more explicit.

Suggested-by: Patrick Bellasi <patrick.bellasi@arm.com>
Signed-off-by: Valentin Schneider <valentin.schneider@arm.com>